moved app to components
