--- src/components/App.js
@@ -1,12 +1,0 @@
-/* eslint-disable react/prefer-stateless-function */
-import React, { Component } from 'react';
-
-export default class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <div>App</div>
-      </div>
-    );
-  }
-}
