+++ src/components/App.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/prefer-stateless-function */
 import React, { Component } from 'react';
-import Display from './components/Display';
-import ButtonPanel from './components/ButtonPanel';
+import Display from './Display';
+import ButtonPanel from './ButtonPanel';
 
 export default class App extends Component {
   constructor() {
